+++ CHANGELOG.md
@@ -29,6 +29,9 @@
 
 ### Added
 
+- [#4708](https://github.com/ChainSafe/forest/issues/4708) Add support for the
+  `Filecoin.EthTraceBlock` RPC method.
+
 ### Changed
 
 ### Removed
@@ -36,6 +39,8 @@
 ### Fixed
 
 - [#5111](https://github.com/ChainSafe/forest/issues/5111) Make F3 work when the node Kademlia is disabled.
+
+- [#5122](https://github.com/ChainSafe/forest/issues/5122) Fix a bug in database garbage collection flow.
 
 ## Forest v.0.23.3 "Plumber"